move utitity function to a separate file
